# src/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Inventory Management Backend API",
    "version": "1.0.0",
    "description": "Production-ready RESTful API for Core Inventory, BOM Product Costing, Explicit Material Batch Allocation, Production Task Execution, Attendance & Overtime, Monthly Payroll, System Notifications Event System, and Reporting/Analytics."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1 Endpoint"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token in the format: Bearer <token>"
      }
    },
...
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/products": {
      "get": {
        "summary": "List / Search Inventory Products & Components",
        "tags": ["Products"],
        "parameters": [
          { "name": "itemType", "in": "query", "schema": { "type": "string", "enum": ["COMPONENT", "PRODUCT"] } },
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          { "name": "pageNo", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "showPerPage", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": { "description": "Paginated list of inventory items" }
        }
      },
      "post": {
        "summary": "Create Item / Component (Admin Only)",
        "tags": ["Products"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "unitPrice", "itemType"],
                "properties": {
                  "name": { "type": "string" },
                  "sku": { "type": "string" },
                  "itemType": { "type": "string", "enum": ["COMPONENT", "PRODUCT"] },
                  "unit": { "type": "string" },
                  "unitPrice": { "type": "number" },
                  "currentStock": { "type": "number" },
                  "lowStockThreshold": { "type": "number" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Product created successfully" }
        }
      }
    },
    "/products/{id}/bom": {
      "get": {
        "summary": "Get Product BOM Tree & Suggested Cost",
        "tags": ["Products"],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "BOM hierarchy and cost breakdown" } }
      },
      "put": {
        "summary": "Replace Product BOM (Admin Only)",
        "tags": ["Products"],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "BOM replaced successfully" } }
      }
    },
...
  }
}